+++ src/main/java/jbyoshi/blockdodge/util/RandomChooser.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package jbyoshi.blockdodge;
+package jbyoshi.blockdodge.util;
 
 import java.util.*;
